refactor: use environment manager service client
