+++ src/services/environment_manager/__init__.py
@@ -1,0 +1,4 @@
+from .environment_manager import EnvironmentManager, FALLBACK_ENV_ID
+from .k8s_environment_manager import EnvironmentManager as KubernetesEnvironmentManager
+
+__all__ = ["EnvironmentManager", "KubernetesEnvironmentManager", "FALLBACK_ENV_ID"]
