+++ tests/create_cube_test_env_for_dev.py
@@ -1,13 +1,13 @@
 
 import asyncio
-from src.services.environment_manager.environment_manager import EnvironmentManager
+from src.services.environment_manager import KubernetesEnvironmentManager
 import logging
 
 logging.basicConfig(level=logging.INFO, format='%(asctime)s - %(name)s - %(levelname)s - %(message)s')
 logger = logging.getLogger(__name__)
 
 async def create_test_env():
-    env_manager = EnvironmentManager()
+    env_manager = KubernetesEnvironmentManager()
     test_env_id = "my-test-env-001"
     
     logger.info(f"Attempting to create environment '{test_env_id}'...")
